+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-## 4.4.0 (unreleased)
+## 5.0.0 (unreleased)
 
 - Fixed error around DST for `group_by_week` with MySQL
 - Fixed error around DST for `group_by_week` and `day_start` with PostgreSQL
 - Fixed bug where `day_start` not applied to `minute_of_hour` with MySQL
 - Fixed bug with non-integer `day_start` and `hour_of_day` with MySQL
+- Removed deprecated positional arguments for time zone and range
 
 ## 4.3.0 (2019-12-26)